+++ client/src/util/functions.ts
@@ -1,0 +1,19 @@
+import { readFileSync } from 'fs';
+
+export function readJSONFile( location: string ) {
+	try {
+		return JSON.parse( readFileSync( location ).toString() );
+	}
+	catch( e ) {
+		console.log( `Problems reading ${ location }: ${ e }` );
+		return {};
+	}
+}
+
+export function pushAll<T>( to: T[], from: T[] ) {
+	if ( from ) {
+		for ( const e of from ) {
+			to.push( e );
+		}
+	}
+}