+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 # Change Log
 
-TODO
+## 0.2.0
+
+* Added the following settings:
+	* mivaIde.MVT.defaultEncodingForVariableConversions
+	* mivaIde.MVT.lint.showWarningOnToolkitUsage
+	* mivaIde.MVT.lint.showWarningOnToolbeltUsage
+	* mivaIde.MVT.lint.showWarningForUnencodedVariables
+* Added some linting rules for MVT.
+* Added Convert & Copy command.
+* Updated flow for selecting MVT:DO function values.